Add optional smallvec support
The `pstoedit` function uses some instances of Vec internally. Generally
these are small, so `smallvec` can be used to keep them on the stack.
For now, a number of five elements on the stack was chosen because this
is the number necessary for `pstoedit -f format input.a output.b`.