Add GitHub Actions CI
Two separate workflows for the two crates. A newer version of Ubuntu is
used because `pstoedit -gstest` as used in the tests does not work out
of the box with pstoedit 3.70 in the CI environment. Only Ubuntu and
MacOS are tested for now as installing pstoedit on Windows is
non-trivial.