Tweak generic type arguments
For `pstoedit_owned` there is only renaming. For `pstoedit_cstr` the
function is made more generic at the cost of ergonomics. However, the
function is private and the flexibility is more important. Before, we
could have `S = CString`, which meant ownership of `argv` was retained
because they were in a slice, but ownership of `gs` would be transferred.
Now, we can have `S = CString` and `T = &CStr` to prevent ownership
transfer of `gs`.